*COsdSetup: show correct option in infobar tuner icon setup item
It could be confused for single stb users to see the option
'ON', although it is not possible.
So it's better to show only a possible or real option at item, if stb is in single mode or
item is disabled,
but the real value of g_settings.infobar_show_tuner is not touched.