Bedit: adapt for infobox
The combination of infobox handling and screen size management doesn't
really fit together. Particular it  could be better implemented some
inheritance, because incorporated functions and variables are
redundant and so the classes can be better coordinated.
Moreover, it doesn't seem sensible to use only a part of the display.
The use of full screen instead of window could simplify a lot.