CNeutrinoEventList: use CComponentsText instead CComponentsInfoBox
add doPaintBg(false), causes no paint or repaint and save of item
background.

CC-Items comes with its own hide(), but it's not very sensefull
to store or paint background here.

NOTE: CComponentsText is more of appropriate for embedded text items
than CComponentsInfoBox. CComponentsInfoBox is more made for hint or
messageboxes.