OPKG-Manager: add opkg manager feature
Signed-off-by: Markus Volk <f_l_k@gmx.net>
Signed-off-by: Thilo Graf <dbt@novatux.de>

Opkg is a lightweight package management system based on Ipkg.

For usage with Neutrino, installed opkg programms opkg-cl and opkg-key
are required and can be find in service update menu,
otherwise this feature is not available.

With this feature it is possible to select packages for install and remove
via gui-menu.

See also:
OPKG-Website
- http://bitbucket.org/opkg/opkg for more details.

Tuxbox:
- http://wiki.tuxbox.org/Neutrino-MP Entwicklungsumgebung#OPKG-Pakete
- http://wiki.tuxbox.org.Neutrino-HD Entwicklungsumgebung:OPKG-SystemInstall

Buildsystems with opkg support:
- http://gitorious.org/neutrino-hd/buildsystem-cs
- http://sourceforge.net/p/bstuff/bs4cs/
- Yocto

OPKG-Manager Neutrino:
- picked from http://gitorious.org/neutrino-mp/martiis-neutrino-mp include shellwindow feature by martii
- update License text
- add missing icon
- use macros to get opkg binaries
- use matching buttons in footer, menu button is already reserved to left menu