neutrino.cpp: add signal OnBeforeRestart
This allows to handle events inside from objects with
matching slots and events without explicit
calls in neutrino.cpp.

Used here general for stopping of thread in timer object.
In some cases it could be necessary to stop timer thread
otherwise it is possible that restart is blocked here.
This happens here automatically, without separate instances
of timer objects.

Conflicts:
	src/neutrino.h