COPKGManager: optimize check size
To calculate the required size for installation here we make a quasi-dry run,
it is a bit awkward, but relatively specific, other solutions are welcome.
We create a temporary test directory and fill it with downloaded or user uploaded package file.
Then we unpack the package and change into temporary testing directory.
The required size results from the size of generated folders and subfolders.
TODO: size of dependencies are not really considered