CPersonalizeGui: try to fix disabled observer item
This should avoid disabling observer items inside personalize menu,
because it's possible to lock out an observer item itself.
In This case it's not possible to
enable this item on runtime. Neutrino restart is required.

This effect was observed in mode_ts if an user has tried to change
settings for tv/radio switch. If all three items were disabled, it was
not possible to enable observed items.

NOTE: I think it's not a good idea to enable certain items in main menu or
full sub menus like settings in mode_ts. ...
What should a user do with that during watch a movie? Play with tuner settings, bouquet-, network-, drive settings etc.?
Some users could cause harm, therefore I can't recommend to hold this current state.