- moviebrowser: add possibility to search for movies in SMS-style
This feature has to be activated first. Menu -> Settings ->
Key Setup -> Edit -> My recordings

In moviebrowser press three times key "3" to select first movie
title starting with a "F" or press two times "6" to select first
movie title starting with a "N".

The new item in keybind-setup is the first step to introduce user-
assignable keys in moviebrowser. So finally the horrible hotkeys
for the cutting functions can be removed and/or replaced.