CComponentsWindow: change order of items
Add header and footer items as first  and  body as last item.
Render of items occurs in listed order. So it's better for performance
while render of window.
This is something more advantageously because all other items are
contained inside body. So we avoid possible delay while rendering
of base items. It looks better on screen.